Improvements to template documentation
There were a couple of situations where using Tumblr\StreamBuilder\Streams\NullStream in our examples was confusing. It didn't make it obvious how much of the heavy lifting in template files comes from using the built-in stream-builder classes.

This updates the documentation to use the fictitious MyApplication\Streams\MyStream to make it obvious which parts of the template file reference our custom code and which parts are using built-in classes